flask/materials-flask-google-login/requirements.txt
click
Flask
Flask-Login
forever
requests
oauthlib
pyOpenSSL
prometheus_client
werkzeug
Pillow